Add UTF-8 support.
While reading manuals on X selections, xt and xaw I have
added support of UTF-8 to xclipboard.

As is seen from the patch I use “UTF8_STRING” atom
in the negotiation of the selection type. There are
other possibilities and it seems that there is no
real standard on this, right? I have checked that
UTF8_STRING works with both GTK+ and QT apps.

In addition to the patch to enable correct display
of international texts in xaw widgets one has to define

*international: true

in one’s XRM.

And enough rich set of bitmap fonts must be also installed.

From bug 16671.

Signed-off-by: James Cloos <cloos@jhcloos.com>